move `MainQueueDispatchDecorator` into separate file
