+++ EssentialFeed/EssentialFeediOS/ImageComments UI/Composers/ImageCommentsUIComposer.swift
@@ -48,29 +48,6 @@ public final class ImageCommentsUIComposer {
 	}
 }
 
-private final class MainQueueDispatchDecorator<T> {
-	private let decoratee: T
-
-	init(decoratee: T) {
-		self.decoratee = decoratee
-	}
-
-	func dispatch(completion: @escaping () -> Void) {
-		guard Thread.isMainThread  else {
-			return DispatchQueue.main.async(execute: completion)
-		}
-		completion()
-	}
-}
-
-extension MainQueueDispatchDecorator: ImageCommentsLoader where T == ImageCommentsLoader {
-	func load(completion: @escaping (ImageCommentsLoader.Result) -> Void) {
-		decoratee.load { [weak self] result in
-			self?.dispatch { completion(result) }
-		}
-	}
-}
-
 private extension ImageCommentsViewController {
 	static func makeWith(delegate: ImageCommentsViewControllerDelegate, title: String) -> ImageCommentsViewController {
 		let bundle = Bundle(for: ImageCommentsViewController.self)
